requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.335
openai==1.3.0
google-generativeai==0.3.1
discord.py==2.3.2
aiohttp==3.8.6
python-multipart==0.0.6
PyMuPDF==1.23.5
python-docx==1.0.1
Pillow==10.1.0
python-dotenv